Backend/requirements.txt
anthropic
yfinance
python-dotenv
flask
flask-cors